+++ ERP-light/src/at/erp/light/view/dto/OutgoingArticleDTO.java
@@ -6,12 +6,12 @@ public class OutgoingArticleDTO {
 	private int outgoingArticleId;
 	private ArticleDTO articleDTO;
 	private Integer articleNr;	// used for order within an incomingDelivery
-	private Double numberpu;
+	private Integer numberpu;
 	
 	public OutgoingArticleDTO(){}
 	
 	public OutgoingArticleDTO(int outgoingArticleId, ArticleDTO articleDTO,
-			Integer articleNr, Double numberpu) {
+			Integer articleNr, Integer numberpu) {
 		super();
 		this.outgoingArticleId = outgoingArticleId;
 		this.articleDTO = articleDTO;
@@ -36,10 +36,10 @@ public class OutgoingArticleDTO {
 	public void setArticleNr(Integer articleNr) {
 		this.articleNr = articleNr;
 	}
-	public Double getNumberpu() {
+	public Integer getNumberpu() {
 		return numberpu;
 	}
-	public void setNumberpu(Double numberpu) {
+	public void setNumberpu(Integer numberpu) {
 		this.numberpu = numberpu;
 	}
